+++ atoms/atom.userInfo.ts
@@ -1,23 +1,8 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { atom } from "recoil";
 
-export const globalState = atom<any>({
-  key: "globalState",
-  default: null,
-});
-
-export const realtimeState = atom<any>({
-  key: "realtimeState",
-  default: null,
-});
-
-export const legendsState = atom<any>({
-  key: "legendsState",
-  default: null,
-});
-
-export const clubState = atom<any>({
-  key: "clubState",
+export const userState = atom<any>({
+  key: "apexUserState",
   default: null,
 });
