--- components/main/stats/index.ts
@@ -1,1 +1,0 @@
-export { default as UserStats } from "./UserStats";
