+++ next.config.js
@@ -1,9 +1,6 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
-  images: {
-    domains: ["secure.download.dm.origin.com", "api.mozambiquehe.re"],
-  },
 };
 
 module.exports = nextConfig;
